fix(getUserApi) - separate it to a different concern!
